Move enforceSingleInstance() to qx-system.h
