+++ comp/core/include/qx/core/qx-system.h
@@ -13,6 +13,8 @@ QString processName(quint32 processId);
 
 bool processIsRunning(QString processName);
 bool processIsRunning(quint32 processID);
+
+bool enforceSingleInstance(QString uniqueAppId);
 }
 
 #endif // QX_SYSTEM_H
