+++ comp/core/src/qx-system.cpp
@@ -37,4 +37,19 @@ bool processIsRunning(QString processName) { return processId(processName); }
  */
 bool processIsRunning(quint32 processID) { return processName(processID).isNull(); }
 
+/*!
+ *  @fn bool enforceSingleInstance(QString uniqueAppId)
+ *
+ *  This function is used to limit a particular application such that only one running instance is
+ *  allowed at one time. Call this function at the point at which you want additional instances to
+ *  terminate, and check the result:
+ *
+ *  If the calling instance is the only one running, the function will return @c true; otherwise
+ *  it returns false.
+ *
+ *  @a uniqueAppId acts as a universal identifier for the application, intended to work across different
+ *  versions and builds, and therefore should be a reasonably unique string that typically is never
+ *  changed in future revisions once set.
+ */
+
 }
